Add endpoint support.
This commit introduces endpoint support for both sql and default node
assignment backend. The SQL backend relies on SQL to get the information about
the patterns that should be used, whereas the default backend relies on
configuration.

The token exchange view now returns an api_endpoint and a user id, but not the
adress of the node anymore (this is not useful for the client).