Pass the bundled certificate to the python worker.
This fixes a misunderstanding of the browserid specification. There are two
things to check when verifying an assertion:

- First, that the bundled certificates are signed with the issuer public cert.
- Second, that the assertion is signed with the bundled cert.

This means that we need to pass the bundled certificate to the crypto worker,
so it can do the verification. What was done previously was to check both the
assertion and the bundled certificate signatures against the issuer
certificate.