Add an error channel to the phose <> crypto worker.
Each protobuf response message now have an error_type and an error field (all
optional). If set, the error will be described in it.

We use that on the tokenserver side to know what's the error in the worker and
return HTTP status code accordingly. Errors in certificate retrieval will lead
to 503 for instance, while other kind of errors (wrong signature) will return a
403.