Don't use the "*" audience for loadtesting assertions.
When loadtesting a server, you now will need to configure the load-testing
script accordingly, meaning that you need to set the audience value to the
audience you have configured on the system under load.

One way to do that is to temporarly accept the "loadtest.local" audience on the
server side and set the audience to "http://loadtest.local" in the load test;
this is what's done in this commit.