Change backend interface and schema for latest protocol changes.
To support Firefox-Accounts-enabled sync, we need to track a
"generation number" and a "client state" for each user, and to
re-assign their uid when the client state changes.  This adds
the necessary interface changes to support that.

On the plus side, we no longer need conditions-accepted so that
has been removed in the shuffle.