Track key-rotation timestamp if provided by the IdP.
For the new "durable" syncstorage backend, we want to be able to
report the user's FxA key id in format "{timestamp}-{key_hash}",
which is the same format used by the FxA "scoped keys" system.
In order to do so, we need to know the timestamp at which the user's
key last changed, which is related to (but not the same as) the
"generation number" that we already get from FxA.

This commit updates tokenserver to accept a new `fxa-keysChangedAt`
field in FxA BrowserID assertions, and to extract an equivalent
value from the `X-KeyID` header provided by OAuth clients. We can
use this in conjunction with the existing `client_state` field
to construct an appropriate key id.