Remove lots of old, incomplete docs files.
This repo contained some documentation that had gone stale, much
that was incomplete, and some that seemed to have been copy-pasted
from a different project and not updated. The only possible outcome
from keeping such docs around is confusion.

So, this commit deletes anything docs-related that is not relevant
to the current structure and operation of the tokenserver. The one
file from the `./docs` directory that made the cut was a list of
available configuration options, which has been moved out of the
sphinx-based docs framework and into a plain standalone .rst file.