Disable some consistency checks on keys_changed_at.
While rolling out the code to track `keys_changed_at`, the system will
be in a state where some servers are writing it to the db and some
servers are not. We can't reliably enforce consistency checks until
we know that all servers are reliably writing it to the db.

Co-Authored-By: Philip Jenvey <pjenvey@underboss.org>