Adjust schemas.py to reflect state of DB after migration scripts.
Commit 71a590f85a27eff6ee4cddfe25aaeb33d210baa2 appears to have introduced
a subtle difference between the schema declared in schemas.py and the one
that would be produced by running all our migrations in order.  It added
a migration making `nodeid` non-nullable and adding an index on `nodeid`,
but updated the schema to make `nodeid` nullable and add an index on the
pair (`service`, `nodeid`).

This commit fixes the schema to match what's in the database in production.
For good measure I have also re-ordered the columns to match the order they
physically appear in the database in production.